rename alphanumeric.py to alphabet_support.py and re-factor respective code
