+++ castervoice/rules/apps/editor/eclipse_rules/eclipse2.py
@@ -3,7 +3,7 @@ from dragonfly import Repeat, Dictation, Function, Choice, MappingRule
 from castervoice.lib.actions import Key
 
 from castervoice.rules.apps.editor.eclipse_rules.eclipse_support import ec_con
-from castervoice.rules.core import alphanumeric
+from castervoice.rules.core import alphabet_support
 from castervoice.lib.ctrl.mgr.rule_details import RuleDetails
 from castervoice.lib.merge.additions import IntegerRefST, Boolean
 from castervoice.lib.merge.state.short import R
@@ -75,9 +75,9 @@ class EclipseRule(MappingRule):
         Dictation("text"),
         Dictation("mim"),
         IntegerRefST("n", 1, 3000),
-        alphanumeric.get_alphabet_choice("a"),
-        alphanumeric.get_alphabet_choice("b"),
-        alphanumeric.get_alphabet_choice("c"),
+        alphabet_support.get_alphabet_choice("a"),
+        alphabet_support.get_alphabet_choice("b"),
+        alphabet_support.get_alphabet_choice("c"),
         Choice("punctuation", {"hash tag": "#"}),
         Boolean("back"),
         Boolean("go"),
