+++ castervoice/rules/core/nav.py
@@ -6,7 +6,7 @@ from dragonfly.actions.action_mimic import Mimic
 from castervoice.lib.actions import Key, Mouse
 from castervoice.rules.core.punctuation_support import double_text_punc_dict, text_punc_dict
 from castervoice.rules.ccr.standard import SymbolSpecs
-from castervoice.rules.core.alphanumeric import caster_alphabet
+from castervoice.rules.core.alphabet_support import caster_alphabet
 from castervoice.lib.const import CCRType
 from castervoice.lib.ctrl.mgr.rule_details import RuleDetails
 from castervoice.lib.merge.additions import IntegerRefST
