+++ castervoice/rules/core/numeric.py
@@ -1,6 +1,6 @@
 from dragonfly import Function
 
-from castervoice.rules.core import alphanumeric
+from castervoice.rules.core import alphabet_support
 from castervoice.lib.const import CCRType
 from castervoice.lib.ctrl.mgr.rule_details import RuleDetails
 from castervoice.lib.merge.additions import IntegerRefST
@@ -12,9 +12,9 @@ class Numbers(MergeRule):
     pronunciation = "numbers"
     mapping = {
         "word number <wn>":
-            R(Function(alphanumeric.word_number, extra="wn")),
+            R(Function(alphabet_support.word_number, extra="wn")),
         "numb <wnKK>":
-            R(Function(alphanumeric.numbers2, extra="wnKK"),
+            R(Function(alphabet_support.numbers2, extra="wnKK"),
               rspec="Number"),
     }
